made checkout button conditional to having items in cart
