--- frontend/app/screens/intro_loading.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { StyleSheet, TextInput, TouchableOpacity, Text, AppRegistry, View, Button, Image } from 'react-native';
-
-class IntroLoading extends React.Component {
-
-  render() {
-    return (
-      <View>
-        
-      </View>
-    );
-  }
-
-}
-
-
-export default IntroLoading
